novo hook para adicionar os filtros aos fretes disponiveis + organização inicial do fluxo de props
